pyproject.toml
[project]
name = "liang-ba"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "akracer==0.0.13",
    "akshare==1.18.21",
    "anyascii==0.3.2",
    "asgiref==3.7.2",
    "asttokens==2.4.1",
    "async-timeout==4.0.3",
    "beautifulsoup4==4.11.2",
    "bs4==0.0.1",
    "certifi==2026.1.4",
    "charset-normalizer==3.3.2",
    "chinesecalendar==1.9.0",
    "coreapi==2.3.3",
    "coreschema==0.0.4",
    "cryptography>=45.0.7",
    "decorator==5.1.1",
    "defusedxml==0.7.1",
    "django==4.2.7",
    "django-ckeditor==6.7.0",
    "django-cors-headers==4.3.1",
    "django-crontab==0.7.1",
    "django-debug-toolbar==4.2.0",
    "django-extensions==3.2.3",
    "django-filter==23.4",
    "django-js-asset==2.1.0",
    "django-modelcluster==6.1",
    "django-permissionedforms==0.1",
    "django-redis==5.4.0",
    "django-taggit==4.0.0",
    "django-treebeard==4.7",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt==5.3.1",
    "draftjs-exporter==2.1.7",
    "et-xmlfile==1.1.0",
    "executing==2.0.1",
    "filetype==1.2.0",
    "html5lib==1.1",
    "idna==3.5",
    "interval==1.0.0",
    "ipython>=8.13.0",
    "itypes==1.2.0",
    "jedi==0.19.1",
    "jinja2==3.1.2",
    "jsonpath==0.82.2",
    "l18n==2021.3",
    "lxml==4.9.3",
    "markdown==3.10.1",
    "markupsafe==2.1.3",
    "matplotlib-inline==0.1.6",
    "psycopg2-binary>=2.9.9",
    "numpy==1.24.4",
    "openpyxl==3.1.2",
    "packaging==23.2",
    "pandas==2.0.3",
    "parso==0.8.3",
    "pbr==6.0.0",
    "pexpect==4.8.0",
    "pickleshare==0.7.5",
    "pillow==10.1.0",
    "pillow-heif==0.13.1",
    "prompt-toolkit==3.0.41",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.2",
    "py-mini-racer==0.6.0",
    "pygments==2.17.2",
    "pyparsing==3.1.1",
    "pypinyin==0.49.0",
    "python-dateutil==2.8.2",
    "pytz==2023.3.post1",
    "redis==5.0.1",
    "requests==2.31.0",
    "simplejson==3.19.2",
    "six==1.16.0",
    "soupsieve==2.5",
    "sqlparse==0.4.4",
    "stack-data==0.6.3",
    "stevedore==5.1.0",
    "tablib==3.5.0",
    "tabulate==0.9.0",
    "telepath==0.3.1",
    "tqdm==4.66.1",
    "traitlets==5.13.0",
    "typing-extensions>=4.11.0,<5",
    "tzdata==2023.3",
    "uritemplate==4.1.1",
    "urllib3==2.1.0",
    "gunicorn>=21.0.0",
    "uwsgi>=2.0.31",
    "wagtail==5.2.1",
    "wcwidth==0.2.12",
    "webencodings==0.5.1",
    "websocket-client==0.57.0",
    "willow==1.6.2",
    "xlrd==2.0.1",
    "xlsxwriter==3.1.9",
    "xlwt==1.3.0",
    "uv",
    "mysql-connector-python>=9.6.0",
]

[[tool.uv.index]]
url = "https://mirrors.huaweicloud.com/repository/pypi/simple"
default = true
